+++ react/index.js
@@ -14,6 +14,7 @@ import ProductQuantityStepper from './components/ProductQuantityStepper'
 import ProductSummaryImage from './components/ProductSummaryImage'
 import ProductSummaryImageLoader from './components/ProductSummaryImageLoader'
 import ProductSummaryPrice from './components/ProductSummaryPrice'
+import ProductSummaryName from './components/ProductSummaryName'
 import displayButtonTypes, {
   getDisplayButtonNames,
   getDisplayButtonValues,
@@ -146,44 +147,10 @@ class ProductSummary extends Component {
     const {
       displayMode,
       product,
-      showBorders,
       name: showFieldsProps,
     } = this.props
 
-    const containerClasses = classNames(
-      'flex items-start',
-      {
-        [`${productSummary.nameContainer} justify-center`]: displayMode !== 'inline',
-        'justify-left w-100 pt5': displayMode === 'inline',
-        'pv5': displayMode === 'small',
-        't-mini pb2': displayMode !== 'normal',
-        'pv6': displayMode === 'normal',
-      }
-    )
-
-    const productName = path(['productName'], product)
-    const skuName = path(['sku', 'name'], product)
-    const brandName = path(['brand'], product)
-
-    const brandNameClasses = classNames('t-body', {
-      't-mini': displayMode === 'small',
-      'c-on-base': displayMode === 'inline',
-    })
-
-    return (
-      <div className={containerClasses}>
-        <ProductName
-          className="overflow-hidden c-on-base"
-          brandNameClass={brandNameClasses}
-          skuNameClass="t-small"
-          loaderClass="pt5 overflow-hidden"
-          name={productName}
-          skuName={skuName}
-          brandName={brandName}
-          {...showFieldsProps}
-        />
-      </div>
-    )
+    return ProductSummaryName({ displayMode, product, showFieldsProps })
   }
 
   render() {
